Add ExceptionHandler tests.
