+++ test/Microsoft.Restier.WebApi.Test/Microsoft.Restier.WebApi.Test.csproj
@@ -93,6 +93,7 @@
     <Compile Include="ExceptionHandlerTests.cs" />
     <Compile Include="ODataDomainControllerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StoreDomain.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
