+++ test/Microsoft.Restier.WebApi.Test/ODataDomainControllerTests.cs
@@ -1,26 +1,12 @@
 ﻿// Copyright (c) Microsoft Corporation.  All rights reserved.
 // Licensed under the MIT License.  See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Linq.Expressions;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 using System.Web.Http;
-using System.Web.OData.Builder;
-using Microsoft.OData.Edm;
-using Microsoft.OData.Edm.Library;
-using Microsoft.Restier.Core;
-using Microsoft.Restier.Core.Model;
-using Microsoft.Restier.Core.Query;
-using Microsoft.Restier.Core.Submit;
 using Xunit;
 
 namespace Microsoft.Restier.WebApi.Test
@@ -164,161 +150,6 @@ namespace Microsoft.Restier.WebApi.Test
             request.Headers.Accept.Add(MediaTypeWithQualityHeaderValue.Parse("application/json"));
             HttpResponseMessage response = await client.SendAsync(request);
             Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-        }
-    }
-
-    public static class StoreModel
-    {
-        public static EdmModel Model { get; private set; }
-
-        public static IEdmEntityType Product { get; private set; }
-
-        static StoreModel()
-        {
-            var builder = new ODataConventionModelBuilder();
-            builder.Namespace = "Microsoft.Restier.WebApi.Test";
-            builder.EntitySet<Product>("Products");
-            builder.Function("GetBestProduct").ReturnsFromEntitySet<Product>("Products");
-            builder.Action("RemoveWorstProduct").ReturnsFromEntitySet<Product>("Products");
-            Model = (EdmModel)builder.GetEdmModel();
-            Product = (IEdmEntityType)Model.FindType("Microsoft.Restier.WebApi.Test.Product");
-        }
-    }
-
-    internal class StoreDomain : DomainBase
-    {
-        protected override DomainConfiguration CreateDomainConfiguration()
-        {
-            var configuration = base.CreateDomainConfiguration();
-            configuration.AddHookHandler<IModelBuilder>(new TestModelProducer(StoreModel.Model));
-            configuration.AddHookHandler<IModelMapper>(new TestModelMapper());
-            configuration.AddHookHandler<IQueryExecutor>(new TestQueryExecutor());
-            configuration.AddHookHandler<IQueryExpressionSourcer>(new TestQueryExpressionSourcer());
-            configuration.AddHookHandler<IChangeSetPreparer>(new TestChangeSetPreparer());
-            configuration.AddHookHandler<ISubmitExecutor>(new TestSubmitExecutor());
-            return configuration;
-        }
-    }
-
-    class Product
-    {
-        public int Id { get; set; }
-
-        public string Name { get; set; }
-
-        [Required]
-        public Address Addr { get; set; }
-
-        public Address Addr2 { get; set; }
-
-        public Address Addr3 { get; set; }
-    }
-
-    class Address
-    {
-        public int Zip { get; set; }
-    }
-
-    class TestModelMapper : IModelMapper
-    {
-        public bool TryGetRelevantType(DomainContext context, string name, out Type relevantType)
-        {
-            relevantType = typeof(Product);
-            return true;
-        }
-
-        public bool TryGetRelevantType(DomainContext context, string namespaceName, string name, out Type relevantType)
-        {
-            relevantType = typeof(Product);
-            return true;
-        }
-    }
-
-    class TestModelProducer : IModelBuilder
-    {
-        private EdmModel model;
-
-        public TestModelProducer(EdmModel model)
-        {
-            this.model = model;
-        }
-
-        public Task<IEdmModel> GetModelAsync(InvocationContext context, CancellationToken cancellationToken)
-        {
-            return Task.FromResult<IEdmModel>(model);
-        }
-    }
-
-    class TestQueryExecutor : IQueryExecutor
-    {
-        public Task<QueryResult> ExecuteQueryAsync<TElement>(
-            QueryContext context,
-            IQueryable<TElement> query,
-            CancellationToken cancellationToken)
-        {
-            return Task.FromResult(new QueryResult(query.ToList()));
-        }
-
-        public Task<QueryResult> ExecuteSingleAsync<TResult>(QueryContext context,
-            IQueryable query,
-            Expression expression,
-            CancellationToken cancellationToken)
-        {
-            return Task.FromResult(new QueryResult(new[] { query.Provider.Execute(expression) }));
-        }
-    }
-
-    class TestQueryExpressionSourcer : IQueryExpressionSourcer
-    {
-        public Expression Source(QueryExpressionContext context, bool embedded)
-        {
-            var a = new[] { new Product
-            {
-                Id = 1,
-                Addr = new Address { Zip = 0001 },
-                Addr2= new Address { Zip = 0002 }
-            } };
-
-            if (!embedded)
-            {
-                if (context.VisitedNode.ToString() == "Source(\"Products\", null)")
-                {
-                    return Expression.Constant(a.AsQueryable());
-                }
-            }
-
-            return context.VisitedNode;
-        }
-    }
-
-    class TestChangeSetPreparer : IChangeSetPreparer
-    {
-        public Task PrepareAsync(SubmitContext context, CancellationToken cancellationToken)
-        {
-            var changeSetEntry = context.ChangeSet.Entries.Single();
-
-            var dataModificationEntry = changeSetEntry as DataModificationEntry;
-            if (dataModificationEntry != null)
-            {
-                dataModificationEntry.Entity = new Product()
-                {
-                    Name = "var1",
-                    Addr = new Address()
-                    {
-                        Zip = 330
-                    }
-                };
-            }
-
-            return Task.FromResult<object>(null);
-        }
-    }
-
-    class TestSubmitExecutor : ISubmitExecutor
-    {
-        public Task<SubmitResult> ExecuteSubmitAsync(SubmitContext context, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(new SubmitResult(context.ChangeSet));
         }
     }
 }
